deploy dialog
