+++ client/components/+deploy/deploy-dialog.scss
@@ -1,3 +1,11 @@
+.DeployDiglog {
+}
+
+.AddDeployDialog{
+  width: 800px;
+  height: 400px;
+}
+
 .top-form{
     padding: 16px;
     border: 1px solid #ccc;
